+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## [Unreleased]
+#### Added
+- Ability to delete key-value
+
+#### Changed
+- Design on key-value page
+
 ## [1.13.0] - 2020-07-16
 
 ## [1.13.0-rc.1] - 2020-06-02